prov/efa: Fix the rdma write with imm error handling for efa_rdm_cq
efa_rdm_cq currently writes the rdma write with imm error to eq
instead of cq. This is wrong because man pages specify that libfabric
should write a cq error with null op_context for completion events
that are not associated with a posted operation, which includes
rdma write with imm data when FI_RX_CQ_DATA mode bit is not required.

Signed-off-by: Jessie Yang <jiaxiyan@amazon.com>